Expand main

+++ pde
@@ -2,47 +2,6 @@
 set -eu
 readonly version='0.1.0'
 
-main() {
-  parse_flags "$@"
-  execute_processing
-}
-
-parse_flags() {
-  while [ "$#" -gt 1 ]; do
-    case "$1" in
-      -h|--help)
-        usage
-        exit 0
-        ;;
-      -v|--version)
-        version
-        exit 0
-        ;;
-      -*)
-        echo "${0##*/}: unrecognized option '$1'" >&2
-        exit 2
-        ;;
-      --)
-        shift
-        break
-        ;;
-      *)
-        break
-        ;;
-    esac
-    shift
-  done
-  if [ "$#" -lt 1 ]; then
-    echo "${0##*/}: no input files" >&2
-    exit 2
-  fi
-  if [ ! -f "$1" ]; then
-    echo "${0##*/}: $1: No such file" >&2
-    exit 2
-  fi
-  src="$1"
-}
-
 usage() {
   cat <<EOF >&2
 Usage: ${0##*/} [OPTION]... SRC
@@ -58,23 +17,54 @@ version() {
   echo "$version" >&2
 }
 
-execute_processing() {
-  workdir="$(mktemp -d "/tmp/${0##*/}.tmp.XXXXXX")"
-  atexit() {
-    rm -f -- "$workdir"
-  }
-  trap 'atexit' EXIT
-  trap 'trap - EXIT; atexit; exit -1' INT PIPE TERM
+while [ "$#" -gt 1 ]; do
+  case "$1" in
+    -h|--help)
+      usage
+      exit 0
+      ;;
+    -v|--version)
+      version
+      exit 0
+      ;;
+    -*)
+      echo "${0##*/}: unrecognized option '$1'" >&2
+      exit 2
+      ;;
+    --)
+      shift
+      break
+      ;;
+    *)
+      break
+      ;;
+  esac
+  shift
+done
+if [ "$#" -lt 1 ]; then
+  echo "${0##*/}: no input files" >&2
+  exit 2
+fi
+if [ ! -f "$1" ]; then
+  echo "${0##*/}: $1: No such file" >&2
+  exit 2
+fi
 
-  mkdir -- "$workdir/sketch"
-  mkdir -- "$workdir/output"
-  cp -- "$src" "$workdir/sketch/sketch.pde"
-  processing-java \
-    --sketch="$workdir/sketch" \
-    --output="$workdir/output" \
-    --force \
-    --run
+src="$1"
+workdir="$(mktemp -d "/tmp/${0##*/}.tmp.XXXXXX")"
+atexit() {
+  rm -f -- "$workdir"
 }
+trap 'atexit' EXIT
+trap 'trap - EXIT; atexit; exit -1' INT PIPE TERM
 
-main "$@"
+mkdir -- "$workdir/sketch"
+mkdir -- "$workdir/output"
+cp -- "$src" "$workdir/sketch/sketch.pde"
+processing-java \
+  --sketch="$workdir/sketch" \
+  --output="$workdir/output" \
+  --force \
+  --run
+
 exit 0
